+++ bridge/engine/EngineMixer.cpp
@@ -1,171 +1,23 @@
 #include "bridge/engine/EngineMixer.h"
 #include "api/DataChannelMessage.h"
-#include "api/DataChannelMessageParser.h"
 #include "bridge/MixerManagerAsync.h"
 #include "bridge/engine/ActiveMediaList.h"
-#include "bridge/engine/AddPacketCacheJob.h"
-#include "bridge/engine/AudioForwarderReceiveJob.h"
-#include "bridge/engine/AudioForwarderRewriteAndSendJob.h"
-#include "bridge/engine/DiscardReceivedVideoPacketJob.h"
-#include "bridge/engine/EncodeJob.h"
 #include "bridge/engine/EngineAudioStream.h"
 #include "bridge/engine/EngineBarbell.h"
 #include "bridge/engine/EngineDataStream.h"
 #include "bridge/engine/EngineStreamDirector.h"
 #include "bridge/engine/EngineVideoStream.h"
-#include "bridge/engine/ProcessMissingVideoPacketsJob.h"
-#include "bridge/engine/SendRtcpJob.h"
-#include "bridge/engine/SetMaxMediaBitrateJob.h"
-#include "bridge/engine/SsrcWhitelist.h"
-#include "bridge/engine/VideoForwarderReceiveJob.h"
-#include "bridge/engine/VideoForwarderRewriteAndSendJob.h"
-#include "bridge/engine/VideoForwarderRtxReceiveJob.h"
 #include "bridge/engine/VideoNackReceiveJob.h"
-#include "codec/Opus.h"
 #include "config/Config.h"
 #include "logger/Logger.h"
 #include "memory/Map.h"
 #include "rtp/RtcpFeedback.h"
-#include "rtp/RtcpHeader.h"
 #include "rtp/RtpHeader.h"
 #include "transport/Transport.h"
-#include "transport/TransportImpl.h"
 #include "utils/SimpleJson.h"
-#include "utils/Span.h"
-#include "webrtc/DataChannel.h"
 #include <cstring>
 
 using namespace bridge;
-
-namespace
-{
-
-const int16_t mixSampleScaleFactor = 4;
-
-memory::UniquePacket createGoodBye(uint32_t ssrc, memory::PacketPoolAllocator& allocator)
-{
-    auto packet = memory::makeUniquePacket(allocator);
-    if (packet)
-    {
-        auto* rtcp = rtp::RtcpGoodbye::create(packet->get(), ssrc);
-        packet->setLength(rtcp->header.size());
-    }
-    return packet;
-}
-
-class RemovePacketCacheJob : public jobmanager::CountedJob
-{
-public:
-    RemovePacketCacheJob(bridge::EngineMixer& mixer,
-        transport::Transport& transport,
-        bridge::SsrcOutboundContext& outboundContext,
-        bridge::MixerManagerAsync& mixerManager)
-        : CountedJob(transport.getJobCounter()),
-          _mixer(mixer),
-          _outboundContext(outboundContext),
-          _mixerManager(mixerManager),
-          _endpointIdHash(transport.getEndpointIdHash())
-    {
-    }
-
-    void run() override
-    {
-        if (_outboundContext.packetCache.isSet() && _outboundContext.packetCache.get())
-        {
-            const auto posted = _mixerManager.asyncFreeVideoPacketCache(_mixer, _outboundContext.ssrc, _endpointIdHash);
-            if (posted)
-            {
-                _outboundContext.packetCache.clear();
-            }
-        }
-    }
-
-private:
-    bridge::EngineMixer& _mixer;
-    bridge::SsrcOutboundContext& _outboundContext;
-    bridge::MixerManagerAsync& _mixerManager;
-    size_t _endpointIdHash;
-};
-
-class FinalizeNonSsrcRewriteOutboundContextJob : public jobmanager::CountedJob
-{
-public:
-    FinalizeNonSsrcRewriteOutboundContextJob(bridge::EngineMixer& mixer,
-        transport::RtcTransport& transport,
-        bridge::SsrcOutboundContext& outboundContext,
-        concurrency::SynchronizationContext& engineSyncContext,
-        bridge::MixerManagerAsync& mixerManager,
-        uint32_t feedbackSsrc)
-        : CountedJob(transport.getJobCounter()),
-          _mixer(mixer),
-          _outboundContext(outboundContext),
-          _transport(transport),
-          _engineSyncContext(engineSyncContext),
-          _mixerManager(mixerManager),
-          _feedbackSsrc(feedbackSsrc)
-    {
-    }
-
-    void run() override
-    {
-        const size_t endpointIdHash = _transport.getEndpointIdHash();
-        const uint32_t ssrc = _outboundContext.ssrc;
-
-        auto packet = createGoodBye(ssrc, _outboundContext.allocator);
-        if (packet)
-        {
-            _transport.protectAndSend(std::move(packet));
-        }
-
-        if (_outboundContext.packetCache.isSet() && _outboundContext.packetCache.get())
-        {
-            // Run RemovePacketCacheJob inside of this job
-            RemovePacketCacheJob(_mixer, _transport, _outboundContext, _mixerManager).run();
-        }
-
-        _transport.removeSrtpLocalSsrc(ssrc);
-
-        _engineSyncContext.post(utils::bind(&EngineMixer::onOutboundContextFinalized,
-            &_mixer,
-            endpointIdHash,
-            ssrc,
-            _feedbackSsrc,
-            _outboundContext.rtpMap.isVideo()));
-    }
-
-private:
-    bridge::EngineMixer& _mixer;
-    bridge::SsrcOutboundContext& _outboundContext;
-    transport::RtcTransport& _transport;
-    concurrency::SynchronizationContext& _engineSyncContext;
-    bridge::MixerManagerAsync& _mixerManager;
-    uint32_t _feedbackSsrc;
-};
-
-/**
- * @return true if this ssrc should be skipped and not forwarded to the videoStream.
- */
-inline bool shouldSkipBecauseOfWhitelist(const bridge::EngineVideoStream& videoStream, const uint32_t ssrc)
-{
-    if (!videoStream.ssrcWhitelist.enabled)
-    {
-        return false;
-    }
-
-    switch (videoStream.ssrcWhitelist.numSsrcs)
-    {
-    case 0:
-        return true;
-    case 1:
-        return videoStream.ssrcWhitelist.ssrcs[0] != ssrc;
-    case 2:
-        return videoStream.ssrcWhitelist.ssrcs[0] != ssrc && videoStream.ssrcWhitelist.ssrcs[1] != ssrc;
-    default:
-        return false;
-    }
-}
-
-} // namespace
 
 namespace bridge
 {
@@ -240,248 +92,15 @@ EngineMixer::EngineMixer(const std::string& id,
 
 EngineMixer::~EngineMixer() {}
 
-void EngineMixer::addAudioStream(EngineAudioStream* engineAudioStream)
+memory::UniquePacket EngineMixer::createGoodBye(uint32_t ssrc, memory::PacketPoolAllocator& allocator)
 {
-    const auto endpointIdHash = engineAudioStream->transport.getEndpointIdHash();
-    if (_engineAudioStreams.contains(endpointIdHash))
+    auto packet = memory::makeUniquePacket(allocator);
+    if (packet)
     {
-        return;
+        auto* rtcp = rtp::RtcpGoodbye::create(packet->get(), ssrc);
+        packet->setLength(rtcp->header.size());
     }
-
-    logger::debug("Add engineAudioStream, transport %s, endpointIdHash %lu, audioMixed %c",
-        _loggableId.c_str(),
-        engineAudioStream->transport.getLoggableId().c_str(),
-        endpointIdHash,
-        engineAudioStream->audioMixed ? 't' : 'f');
-
-    _engineAudioStreams.emplace(endpointIdHash, engineAudioStream);
-    if (engineAudioStream->audioMixed)
-    {
-        _numMixedAudioStreams++;
-    }
-
-    auto neighbourIt = _neighbourMemberships.emplace(endpointIdHash, endpointIdHash);
-    if (neighbourIt.second)
-    {
-        auto& neighbourList = neighbourIt.first->second.memberships;
-        for (auto& it : engineAudioStream->neighbours)
-        {
-            neighbourList.push_back(it.first);
-        }
-    }
-    else
-    {
-        logger::error("Failed to setup neighbour list for audio stream %zu", _loggableId.c_str(), endpointIdHash);
-    }
-
-    const auto mapRevision = _activeMediaList->getMapRevision();
-    _activeMediaList->addAudioParticipant(endpointIdHash, engineAudioStream->endpointId.c_str());
-    if (mapRevision != _activeMediaList->getMapRevision())
-    {
-        sendUserMediaMapMessageToAll();
-        sendUserMediaMapMessageOverBarbells();
-    }
-    updateBandwidthFloor();
-
-    sendAudioStreamToRecording(*engineAudioStream, true);
-}
-
-void EngineMixer::removeStream(const EngineAudioStream* engineAudioStream)
-{
-    const auto endpointIdHash = engineAudioStream->endpointIdHash;
-
-    const auto mapRevision = _activeMediaList->getMapRevision();
-    _activeMediaList->removeAudioParticipant(endpointIdHash);
-
-    if (mapRevision != _activeMediaList->getMapRevision())
-    {
-        sendUserMediaMapMessageToAll();
-        sendUserMediaMapMessageOverBarbells();
-    }
-
-    updateBandwidthFloor();
-
-    if (engineAudioStream->transport.isConnected())
-    {
-        // Job count will increase delaying the deletion of transport by Mixer.
-        auto* context = engineAudioStream->ssrcOutboundContexts.getItem(engineAudioStream->localSsrc);
-        if (context)
-        {
-            auto goodByePacket = createGoodBye(context->ssrc, context->allocator);
-            if (goodByePacket)
-            {
-                engineAudioStream->transport.getJobQueue().addJob<SendRtcpJob>(std::move(goodByePacket),
-                    engineAudioStream->transport);
-            }
-        }
-    }
-
-    if (engineAudioStream->audioMixed)
-    {
-        _numMixedAudioStreams--;
-    }
-
-    logger::debug("Remove engineAudioStream, transport %s",
-        _loggableId.c_str(),
-        engineAudioStream->transport.getLoggableId().c_str());
-
-    if (engineAudioStream->remoteSsrc.isSet())
-    {
-        decommissionInboundContext(engineAudioStream->remoteSsrc.get());
-        _mixerSsrcAudioBuffers.erase(engineAudioStream->remoteSsrc.get());
-
-        markAssociatedAudioOutboundContextsForDeletion(engineAudioStream);
-        sendAudioStreamToRecording(*engineAudioStream, false);
-    }
-
-    _engineAudioStreams.erase(endpointIdHash);
-
-    engineAudioStream->transport.postOnQueue(
-        [this, engineAudioStream]() { _messageListener.asyncAudioStreamRemoved(*this, *engineAudioStream); });
-}
-
-void EngineMixer::addVideoStream(EngineVideoStream* engineVideoStream)
-{
-    const auto endpointIdHash = engineVideoStream->endpointIdHash;
-    if (_engineVideoStreams.contains(endpointIdHash))
-    {
-        return;
-    }
-
-    logger::debug("Add engineVideoStream, transport %s, endpointIdHash %lu",
-        _loggableId.c_str(),
-        engineVideoStream->transport.getLoggableId().c_str(),
-        endpointIdHash);
-
-    if (_probingVideoStreams)
-    {
-        startProbingVideoStream(*engineVideoStream);
-    }
-    const auto mapRevision = _activeMediaList->getMapRevision();
-    const auto it = _engineVideoStreams.emplace(endpointIdHash, engineVideoStream);
-    if (!it.second)
-    {
-        logger::error("Emplace video stream has failed, transport %s, endpointIdHash %lu",
-            _loggableId.c_str(),
-            engineVideoStream->transport.getLoggableId().c_str(),
-            endpointIdHash);
-    }
-
-    if (engineVideoStream->simulcastStream.numLevels > 0)
-    {
-        _engineStreamDirector->addParticipant(endpointIdHash, engineVideoStream->simulcastStream);
-        _activeMediaList->addVideoParticipant(endpointIdHash,
-            engineVideoStream->simulcastStream,
-            engineVideoStream->secondarySimulcastStream,
-            engineVideoStream->endpointId.c_str());
-    }
-    else
-    {
-        _engineStreamDirector->addParticipant(endpointIdHash);
-    }
-    updateBandwidthFloor();
-    sendLastNListMessageToAll();
-    if (mapRevision != _activeMediaList->getMapRevision())
-    {
-        sendUserMediaMapMessageToAll();
-        sendUserMediaMapMessageOverBarbells();
-    }
-
-    sendVideoStreamToRecording(*engineVideoStream, true);
-}
-
-void EngineMixer::removeStream(const EngineVideoStream* engineVideoStream)
-{
-    engineVideoStream->transport.getJobQueue().getJobManager().abortTimedJob(engineVideoStream->transport.getId(),
-        engineVideoStream->localSsrc);
-
-    auto* outboundContext = engineVideoStream->ssrcOutboundContexts.getItem(engineVideoStream->localSsrc);
-    if (outboundContext)
-    {
-        outboundContext->markedForDeletion = true;
-        if (engineVideoStream->transport.isConnected())
-        {
-            stopProbingVideoStream(*engineVideoStream);
-            auto goodByePacket = createGoodBye(outboundContext->ssrc, outboundContext->allocator);
-            if (goodByePacket)
-            {
-                engineVideoStream->transport.getJobQueue().addJob<SendRtcpJob>(std::move(goodByePacket),
-                    engineVideoStream->transport);
-            }
-        }
-    }
-
-    for (auto& ssrcOutboundContextPair : engineVideoStream->ssrcOutboundContexts)
-    {
-        if (!ssrcOutboundContextPair.second.markedForDeletion)
-        {
-            engineVideoStream->transport.getJobQueue().addJob<RemovePacketCacheJob>(*this,
-                engineVideoStream->transport,
-                ssrcOutboundContextPair.second,
-                _messageListener);
-        }
-    }
-
-    if (engineVideoStream->simulcastStream.numLevels != 0)
-    {
-        for (auto& simulcastLevel : engineVideoStream->simulcastStream.getLevels())
-        {
-            decommissionInboundContext(simulcastLevel.ssrc);
-            decommissionInboundContext(simulcastLevel.feedbackSsrc);
-        }
-
-        markAssociatedVideoOutboundContextsForDeletion(engineVideoStream,
-            engineVideoStream->simulcastStream.levels[0].ssrc,
-            engineVideoStream->simulcastStream.levels[0].feedbackSsrc);
-    }
-
-    if (engineVideoStream->secondarySimulcastStream.isSet() &&
-        engineVideoStream->secondarySimulcastStream.get().numLevels != 0)
-    {
-        for (auto& simulcastLevel : engineVideoStream->secondarySimulcastStream.get().getLevels())
-        {
-            decommissionInboundContext(simulcastLevel.ssrc);
-            decommissionInboundContext(simulcastLevel.feedbackSsrc);
-        }
-
-        markAssociatedVideoOutboundContextsForDeletion(engineVideoStream,
-            engineVideoStream->secondarySimulcastStream.get().levels[0].ssrc,
-            engineVideoStream->secondarySimulcastStream.get().levels[0].feedbackSsrc);
-    }
-
-    const auto endpointIdHash = engineVideoStream->endpointIdHash;
-
-    const auto mapRevision = _activeMediaList->getMapRevision();
-    _activeMediaList->removeVideoParticipant(endpointIdHash);
-
-    if (mapRevision != _activeMediaList->getMapRevision())
-    {
-        sendUserMediaMapMessageToAll();
-        sendUserMediaMapMessageOverBarbells();
-    }
-
-    _engineStreamDirector->removeParticipant(endpointIdHash);
-    _engineStreamDirector->removeParticipantPins(endpointIdHash);
-    updateBandwidthFloor();
-
-    sendVideoStreamToRecording(*engineVideoStream, false);
-
-    logger::debug("Remove engineVideoStream, transport %s, endpointIdHash %lu",
-        _loggableId.c_str(),
-        engineVideoStream->transport.getLoggableId().c_str(),
-        endpointIdHash);
-
-    const bool streamFound = _engineVideoStreams.erase(endpointIdHash);
-    if (!streamFound)
-    {
-        logger::error("engineVideoStream has not been found, transport %s, endpointIdHash %lu",
-            _loggableId.c_str(),
-            engineVideoStream->transport.getLoggableId().c_str(),
-            endpointIdHash);
-    }
-
-    engineVideoStream->transport.postOnQueue(
-        [this, engineVideoStream]() { _messageListener.asyncVideoStreamRemoved(*this, *engineVideoStream); });
+    return packet;
 }
 
 void EngineMixer::addDataSteam(EngineDataStream* engineDataStream)
@@ -524,205 +143,6 @@ void EngineMixer::startTransport(transport::RtcTransport& transport)
     // race condition and sync problems with ice session and srtpclient
     transport.start();
     transport.connect();
-}
-
-void EngineMixer::reconfigureAudioStream(const transport::RtcTransport& transport, const uint32_t remoteSsrc)
-{
-    auto* engineAudioStream = _engineAudioStreams.getItem(transport.getEndpointIdHash());
-    if (!engineAudioStream)
-    {
-        return;
-    }
-
-    if (engineAudioStream->remoteSsrc.isSet() && engineAudioStream->remoteSsrc.get() != remoteSsrc)
-    {
-        decommissionInboundContext(engineAudioStream->remoteSsrc.get());
-        sendAudioStreamToRecording(*engineAudioStream, false);
-    }
-
-    if (remoteSsrc != 0)
-    {
-        engineAudioStream->remoteSsrc.set(remoteSsrc);
-        sendAudioStreamToRecording(*engineAudioStream, true);
-    }
-    else
-    {
-        engineAudioStream->remoteSsrc = utils::Optional<uint32_t>();
-    }
-    updateBandwidthFloor();
-}
-
-void EngineMixer::reconfigureVideoStream(const transport::RtcTransport& transport,
-    const SsrcWhitelist& ssrcWhitelist,
-    const SimulcastStream& simulcastStream,
-    const utils::Optional<SimulcastStream>& secondarySimulcastStream)
-{
-    const auto endpointIdHash = transport.getEndpointIdHash();
-
-    auto* engineVideoStream = _engineVideoStreams.getItem(endpointIdHash);
-    if (!engineVideoStream)
-    {
-        return;
-    }
-
-    logger::debug("Reconfigure %s video stream, endpointIdHash %lu, whitelist %c %u %u %u",
-        _loggableId.c_str(),
-        secondarySimulcastStream.isSet() ? "secondary" : "primary",
-        transport.getEndpointIdHash(),
-        ssrcWhitelist.enabled ? 't' : 'f',
-        ssrcWhitelist.numSsrcs,
-        ssrcWhitelist.ssrcs[0],
-        ssrcWhitelist.ssrcs[1]);
-
-    memory::Array<uint32_t, 12> decommissionedSsrcs;
-
-    const auto mapRevision = _activeMediaList->getMapRevision();
-    if (engineVideoStream->simulcastStream.numLevels != 0)
-    {
-        const bool hasLevelZeroChanged = simulcastStream.numLevels == 0 ||
-            simulcastStream.levels[0].ssrc != engineVideoStream->simulcastStream.levels[0].ssrc;
-
-        if (hasLevelZeroChanged)
-        {
-            removeVideoSsrcFromRecording(*engineVideoStream, engineVideoStream->simulcastStream.levels[0].ssrc);
-        }
-
-        for (auto& simulcastLevel : engineVideoStream->simulcastStream.getLevels())
-        {
-            decommissionedSsrcs.push_back(simulcastLevel.ssrc);
-            decommissionedSsrcs.push_back(simulcastLevel.feedbackSsrc);
-        }
-    }
-
-    if ((engineVideoStream->secondarySimulcastStream.isSet() &&
-            engineVideoStream->secondarySimulcastStream.get().numLevels != 0))
-    {
-        const bool hasLevelZeroChanged = !secondarySimulcastStream.isSet() ||
-            secondarySimulcastStream.get().levels[0].ssrc !=
-                engineVideoStream->secondarySimulcastStream.get().levels[0].ssrc;
-
-        if (hasLevelZeroChanged)
-        {
-            removeVideoSsrcFromRecording(*engineVideoStream,
-                engineVideoStream->secondarySimulcastStream.get().levels[0].ssrc);
-        }
-
-        for (auto& simulcastLevel : engineVideoStream->secondarySimulcastStream.get().getLevels())
-        {
-            decommissionedSsrcs.push_back(simulcastLevel.ssrc);
-            decommissionedSsrcs.push_back(simulcastLevel.feedbackSsrc);
-        }
-    }
-
-    engineVideoStream->simulcastStream = simulcastStream;
-    engineVideoStream->secondarySimulcastStream = secondarySimulcastStream;
-
-    _engineStreamDirector->removeParticipant(endpointIdHash);
-    _activeMediaList->removeVideoParticipant(endpointIdHash);
-
-    auto decommissionedSsrcsEndIt = decommissionedSsrcs.end();
-
-    if (engineVideoStream->simulcastStream.numLevels > 0)
-    {
-        if (secondarySimulcastStream.isSet() && secondarySimulcastStream.get().numLevels > 0)
-        {
-            engineVideoStream->secondarySimulcastStream = secondarySimulcastStream;
-            _engineStreamDirector->addParticipant(endpointIdHash,
-                engineVideoStream->simulcastStream,
-                &engineVideoStream->secondarySimulcastStream.get());
-
-            for (auto& simulcastLevel : engineVideoStream->secondarySimulcastStream.get().getLevels())
-            {
-                decommissionedSsrcsEndIt =
-                    std::remove(decommissionedSsrcs.begin(), decommissionedSsrcsEndIt, simulcastLevel.ssrc);
-                decommissionedSsrcsEndIt =
-                    std::remove(decommissionedSsrcs.begin(), decommissionedSsrcsEndIt, simulcastLevel.feedbackSsrc);
-            }
-        }
-        else
-        {
-            _engineStreamDirector->addParticipant(endpointIdHash, engineVideoStream->simulcastStream);
-        }
-
-        for (auto& simulcastLevel : engineVideoStream->simulcastStream.getLevels())
-        {
-            decommissionedSsrcsEndIt =
-                std::remove(decommissionedSsrcs.begin(), decommissionedSsrcsEndIt, simulcastLevel.ssrc);
-            decommissionedSsrcsEndIt =
-                std::remove(decommissionedSsrcs.begin(), decommissionedSsrcsEndIt, simulcastLevel.feedbackSsrc);
-        }
-
-        _activeMediaList->addVideoParticipant(endpointIdHash,
-            engineVideoStream->simulcastStream,
-            engineVideoStream->secondarySimulcastStream,
-            engineVideoStream->endpointId.c_str());
-
-        restoreDirectorStreamActiveState(*engineVideoStream, engineVideoStream->simulcastStream);
-        if (secondarySimulcastStream.isSet() && secondarySimulcastStream.get().numLevels > 0)
-        {
-            restoreDirectorStreamActiveState(*engineVideoStream, secondarySimulcastStream.get());
-        }
-    }
-    else
-    {
-        _engineStreamDirector->addParticipant(endpointIdHash);
-    }
-
-    const auto decommissionedSpan = utils::Span<uint32_t>(decommissionedSsrcs.begin(), decommissionedSsrcsEndIt);
-    for (const auto& decommissionedSsrc : decommissionedSpan)
-    {
-        decommissionInboundContext(decommissionedSsrc);
-    }
-
-    updateBandwidthFloor();
-    sendLastNListMessageToAll();
-    if (mapRevision != _activeMediaList->getMapRevision())
-    {
-        sendUserMediaMapMessageToAll();
-        sendUserMediaMapMessageOverBarbells();
-    }
-    sendVideoStreamToRecording(*engineVideoStream, true);
-
-    engineVideoStream->ssrcWhitelist = ssrcWhitelist;
-}
-
-void EngineMixer::addVideoPacketCache(const uint32_t ssrc, const size_t endpointIdHash, PacketCache* videoPacketCache)
-{
-    transport::RtcTransport* transport = nullptr;
-    bridge::SsrcOutboundContext* ssrcContext = nullptr;
-    auto* videoStream = _engineVideoStreams.getItem(endpointIdHash);
-    if (videoStream)
-    {
-        ssrcContext = videoStream->ssrcOutboundContexts.getItem(ssrc);
-        transport = &videoStream->transport;
-    }
-    else
-    {
-        auto* barbell = _engineBarbells.getItem(endpointIdHash);
-        if (barbell)
-        {
-            ssrcContext = barbell->ssrcOutboundContexts.getItem(ssrc);
-            transport = &barbell->transport;
-        }
-    }
-
-    if (ssrcContext && transport)
-    {
-        if (!transport->getJobQueue().addJob<AddPacketCacheJob>(*transport, *ssrcContext, videoPacketCache))
-        {
-            logger::warn("JobQueue full. Failed to add packet cache to %zu, ssrc %u, %s",
-                _loggableId.c_str(),
-                endpointIdHash,
-                ssrc,
-                transport->getLoggableId().c_str());
-        }
-    }
-}
-
-void EngineMixer::addAudioBuffer(const uint32_t ssrc, AudioBuffer* audioBuffer)
-{
-    _mixerSsrcAudioBuffers.erase(ssrc);
-    _mixerSsrcAudioBuffers.emplace(ssrc, audioBuffer);
 }
 
 void EngineMixer::clear()
@@ -782,29 +202,6 @@ void EngineMixer::run(const uint64_t engineIterationStartTimestamp)
 
     // 6. Maintain transports.
     runTransportTicks(engineIterationStartTimestamp);
-}
-
-void EngineMixer::processMissingPackets(const uint64_t timestamp)
-{
-    for (auto& ssrcInboundContextEntry : _ssrcInboundContexts)
-    {
-        auto& ssrcInboundContext = *ssrcInboundContextEntry.second;
-        if (ssrcInboundContext.rtpMap.format != RtpMap::Format::VP8)
-        {
-            continue;
-        }
-
-        if (EngineBarbell::isFromBarbell(ssrcInboundContext.sender->getTag()))
-        {
-            processBarbellMissingPackets(ssrcInboundContext);
-        }
-        else
-        {
-            processEngineMissingPackets(ssrcInboundContext);
-        }
-    }
-
-    processRecordingUnackedPackets(timestamp);
 }
 
 void EngineMixer::runDominantSpeakerCheck(const uint64_t engineIterationStartTimestamp)
@@ -1084,127 +481,6 @@ EngineStats::MixerStats EngineMixer::gatherStats(const uint64_t iterationStartTi
     return stats;
 }
 
-void EngineMixer::onAudioRtpPacketReceived(SsrcInboundContext& ssrcContext,
-    transport::RtcTransport* sender,
-    memory::UniquePacket packet,
-    const uint32_t extendedSequenceNumber,
-    const uint64_t timestamp)
-{
-    // TODO we should also check if this ssrc is in use, mixed clients are present, or it needs to be forwarded over
-    // barbell. If not, we can drop this packet.
-    if (!_engineAudioStreams.empty())
-    {
-        sender->getJobQueue().addJob<bridge::AudioForwarderReceiveJob>(std::move(packet),
-            sender,
-            *this,
-            ssrcContext,
-            *_activeMediaList,
-            _config.audio.silenceThresholdLevel,
-            _numMixedAudioStreams != 0,
-            !_engineBarbells.empty() || _engineAudioStreams.size() > 2,
-            extendedSequenceNumber);
-    }
-}
-
-void EngineMixer::onVideoRtpPacketReceived(SsrcInboundContext& ssrcContext,
-    transport::RtcTransport* sender,
-    memory::UniquePacket packet,
-    const uint32_t extendedSequenceNumber,
-    const uint64_t timestamp)
-{
-
-    if (ssrcContext.shouldDropPackets)
-    {
-        return;
-    }
-
-    const bool isFromBarbell = EngineBarbell::isFromBarbell(sender->getTag());
-    const size_t endpointIdHash = ssrcContext.endpointIdHash;
-
-    auto* videoStream = _engineVideoStreams.getItem(endpointIdHash);
-    if (!videoStream && !isFromBarbell)
-    {
-        return;
-    }
-
-    const auto isSenderInLastNList = _activeMediaList->isInActiveVideoList(endpointIdHash);
-    const bool mustBeForwardedOnBarbells = isSenderInLastNList && !_engineBarbells.empty() && !isFromBarbell;
-
-    if (!mustBeForwardedOnBarbells && !ssrcContext.isSsrcUsed.load())
-    {
-        sender->getJobQueue().addJob<bridge::DiscardReceivedVideoPacketJob>(std::move(packet),
-            sender,
-            ssrcContext,
-            extendedSequenceNumber);
-        return;
-    }
-
-    sender->getJobQueue().addJob<bridge::VideoForwarderReceiveJob>(std::move(packet),
-        _sendAllocator,
-        sender,
-        *this,
-        ssrcContext,
-        _localVideoSsrc,
-        extendedSequenceNumber,
-        timestamp);
-}
-
-void EngineMixer::onVideoRtpRtxPacketReceived(SsrcInboundContext& rtxSsrcContext,
-    transport::RtcTransport* sender,
-    memory::UniquePacket packet,
-    const uint32_t extendedSequenceNumber,
-    const uint64_t timestamp)
-{
-    const bool isFromBarbell = EngineBarbell::isFromBarbell(sender->getTag());
-    const size_t endpointIdHash = rtxSsrcContext.endpointIdHash;
-    EngineVideoStream* videoStream = _engineVideoStreams.getItem(endpointIdHash);
-    if (!videoStream && !isFromBarbell)
-    {
-        return;
-    }
-
-    uint32_t mainSsrc;
-    const uint32_t rtxSsrc = rtxSsrcContext.ssrc;
-    if (videoStream && !_engineStreamDirector->getSsrc(videoStream->endpointIdHash, rtxSsrc, mainSsrc))
-    {
-        return;
-    }
-    else if (isFromBarbell)
-    {
-        auto ssrc = findBarbellMainSsrc(sender->getEndpointIdHash(), rtxSsrc);
-        if (!ssrc.isSet())
-        {
-            return;
-        }
-        mainSsrc = ssrc.get();
-    }
-
-    auto mainSsrcContext = _ssrcInboundContexts.getItem(mainSsrc);
-    if (!mainSsrcContext || mainSsrcContext->shouldDropPackets)
-    {
-        return;
-    }
-
-    mainSsrcContext->onRtpPacketReceived(timestamp);
-
-    const auto isSenderInLastNList = _activeMediaList->isInActiveVideoList(endpointIdHash);
-    const bool mustBeForwardedOnBarbells = isSenderInLastNList && !_engineBarbells.empty() && !isFromBarbell;
-
-    // Optimization with isSsrcUsed is not need for barbell (barbell sends only necessary streams).
-    if (videoStream && !mustBeForwardedOnBarbells && !mainSsrcContext->isSsrcUsed)
-    {
-        return;
-    }
-
-    sender->getJobQueue().addJob<bridge::VideoForwarderRtxReceiveJob>(std::move(packet),
-        sender,
-        *this,
-        rtxSsrcContext,
-        *mainSsrcContext,
-        mainSsrc,
-        extendedSequenceNumber);
-}
-
 void EngineMixer::onConnected(transport::RtcTransport* sender)
 {
     logger::debug("transport connected", sender->getLoggableId().c_str());
@@ -1244,43 +520,6 @@ void EngineMixer::handleSctpControl(const size_t endpointIdHash, memory::UniqueP
             sendUserMediaMapMessage(endpointIdHash);
         }
     }
-}
-
-void EngineMixer::pinEndpoint(const size_t endpointIdHash, const size_t targetEndpointIdHash)
-{
-    logger::debug("pinEndpoint %lu -> %lu", _loggableId.c_str(), endpointIdHash, targetEndpointIdHash);
-    auto oldTarget = _engineStreamDirector->pin(endpointIdHash, targetEndpointIdHash);
-    if (oldTarget == targetEndpointIdHash)
-    {
-        return;
-    }
-
-    auto* videoStream = _engineVideoStreams.getItem(endpointIdHash);
-    if (videoStream)
-    {
-        const auto oldPinSsrc = videoStream->pinSsrc;
-        if (oldPinSsrc.isSet())
-        {
-            videoStream->videoPinSsrcs.push(oldPinSsrc.get());
-            videoStream->pinSsrc.clear();
-        }
-        if (targetEndpointIdHash)
-        {
-            SimulcastLevel newPinSsrc;
-            if (videoStream->videoPinSsrcs.pop(newPinSsrc))
-            {
-                videoStream->pinSsrc.set(newPinSsrc);
-                logger::debug("EndpointIdHash %zu pin %zu, pin ssrc %u",
-                    _loggableId.c_str(),
-                    endpointIdHash,
-                    targetEndpointIdHash,
-                    newPinSsrc.ssrc);
-            }
-        }
-    }
-
-    sendLastNListMessage(endpointIdHash);
-    sendUserMediaMapMessage(endpointIdHash);
 }
 
 void EngineMixer::sendEndpointMessage(const size_t toEndpointIdHash,
@@ -1451,42 +690,6 @@ void EngineMixer::onRtpPacketReceived(transport::RtcTransport* sender,
             getLoggableId().c_str(),
             rtpHeader->payloadType);
         break;
-    }
-}
-
-void EngineMixer::onForwarderAudioRtpPacketDecrypted(SsrcInboundContext& inboundContext,
-    memory::UniquePacket packet,
-    const uint32_t extendedSequenceNumber)
-{
-    assert(packet);
-    if (!_incomingForwarderAudioRtp.push(
-            IncomingPacketInfo(std::move(packet), &inboundContext, extendedSequenceNumber)))
-    {
-        logger::error("Failed to push incoming forwarder audio packet onto queue", getLoggableId().c_str());
-        assert(false);
-    }
-}
-
-void EngineMixer::onForwarderVideoRtpPacketDecrypted(SsrcInboundContext& inboundContext,
-    memory::UniquePacket packet,
-    const uint32_t extendedSequenceNumber)
-{
-    assert(packet);
-    if (!_incomingForwarderVideoRtp.push(
-            IncomingPacketInfo(std::move(packet), &inboundContext, extendedSequenceNumber)))
-    {
-        logger::error("Failed to push incoming forwarder video packet onto queue", getLoggableId().c_str());
-        assert(false);
-    }
-}
-
-void EngineMixer::onMixerAudioRtpPacketDecoded(SsrcInboundContext& inboundContext, memory::UniqueAudioPacket packet)
-{
-    assert(packet);
-    if (!_incomingMixerAudioRtp.push(IncomingAudioPacketInfo(std::move(packet), &inboundContext)))
-    {
-        logger::error("Failed to push incoming mixer audio packet onto queue", getLoggableId().c_str());
-        assert(false);
     }
 }
 
@@ -1762,189 +965,6 @@ SsrcInboundContext* EngineMixer::emplaceInboundSsrcContext(const uint32_t ssrc,
     return nullptr;
 }
 
-namespace
-{
-template <class V, class T>
-bool isNeighbour(const V& groupList, const T& lookupTable)
-{
-    for (auto& entry : groupList)
-    {
-        if (lookupTable.contains(entry))
-        {
-            return true;
-        }
-    }
-    return false;
-}
-
-template <class TMap>
-bool areNeighbours(const TMap& table1, const TMap& table2)
-{
-    if (table1.size() < table2.size())
-    {
-        for (auto& entry : table1)
-        {
-            if (table2.contains(entry.first))
-            {
-                return true;
-            }
-        }
-    }
-    else
-    {
-        for (auto& entry : table2)
-        {
-            if (table1.contains(entry.first))
-            {
-                return true;
-            }
-        }
-    }
-
-    return false;
-}
-
-} // namespace
-
-void EngineMixer::forwardAudioRtpPacket(IncomingPacketInfo& packetInfo, uint64_t timestamp)
-{
-    const auto* rtpHeader = rtp::RtpHeader::fromPacket(*packetInfo.packet());
-    auto srcUserId = getC9UserId(rtpHeader->ssrc);
-
-    const auto& audioSsrcRewriteMap = _activeMediaList->getAudioSsrcRewriteMap();
-    const auto rewriteMapItr = audioSsrcRewriteMap.find(packetInfo.packet()->endpointIdHash);
-    const bool sourceMapped = (rewriteMapItr != audioSsrcRewriteMap.end());
-    const auto originalSsrc = packetInfo.inboundContext()->ssrc;
-
-    for (auto& audioStreamEntry : _engineAudioStreams)
-    {
-        auto* audioStream = audioStreamEntry.second;
-
-        if (audioStream && &audioStream->transport == packetInfo.transport())
-        {
-            if (!audioStream->detectedAudioSsrc.isSet())
-            {
-                logger::info("%zu detected audio ssrc %u, negotiated ssrc %u, audio-lvl-extid %u",
-                    _loggableId.c_str(),
-                    audioStream->endpointIdHash,
-                    originalSsrc,
-                    audioStream->remoteSsrc.get(),
-                    audioStream->rtpMap.audioLevelExtId.valueOr(0));
-            }
-            audioStream->detectedAudioSsrc.set(originalSsrc);
-        }
-        if (!audioStream || &audioStream->transport == packetInfo.transport() || audioStream->audioMixed)
-        {
-            continue;
-        }
-        if (audioStream->ssrcRewrite && !sourceMapped)
-        {
-            continue;
-        }
-
-        if (srcUserId.isSet() && audioStream->neighbours.contains(srcUserId.get()))
-        {
-            continue;
-        }
-
-        if (!audioStream->neighbours.empty())
-        {
-            auto* srcMemberships = _neighbourMemberships.getItem(packetInfo.packet()->endpointIdHash);
-            if (srcMemberships && isNeighbour(srcMemberships->memberships, audioStream->neighbours))
-            {
-                continue;
-            }
-        }
-
-        SsrcOutboundContext* ssrcOutboundContext = nullptr;
-        if (audioStream->ssrcRewrite)
-        {
-            ssrcOutboundContext = obtainOutboundSsrcContext(audioStream->endpointIdHash,
-                audioStream->ssrcOutboundContexts,
-                rewriteMapItr->second,
-                audioStream->rtpMap);
-        }
-        else
-        {
-            ssrcOutboundContext = obtainOutboundForwardSsrcContext(audioStream->endpointIdHash,
-                audioStream->ssrcOutboundContexts,
-                originalSsrc,
-                audioStream->rtpMap);
-        }
-
-        if (!ssrcOutboundContext)
-        {
-            continue;
-        }
-
-        auto packet = memory::makeUniquePacket(_sendAllocator, *packetInfo.packet());
-        if (packet)
-        {
-            audioStream->transport.getJobQueue().addJob<AudioForwarderRewriteAndSendJob>(*ssrcOutboundContext,
-                *(packetInfo.inboundContext()),
-                std::move(packet),
-                packetInfo.extendedSequenceNumber(),
-                audioStream->transport);
-        }
-        else
-        {
-            logger::warn("send allocator depleted. forwardAudioRtpPacket", _loggableId.c_str());
-        }
-    }
-}
-
-/**
- * Append RTP audio to pre-buffer for this ssrc in _mixerSsrcAudioBuffers
- */
-void EngineMixer::addPacketToMixerBuffers(const IncomingAudioPacketInfo& packetInfo,
-    const uint64_t timestamp,
-    bool overrunLogSpamGuard)
-{
-    const auto rtpHeader = rtp::RtpHeader::fromPacket(*packetInfo.packet());
-    if (!rtpHeader)
-    {
-        return;
-    }
-
-    const auto ssrc = rtpHeader->ssrc;
-    const auto sequenceNumber = rtpHeader->sequenceNumber;
-    const auto payloadStart = rtpHeader->getPayload();
-    const uint32_t payloadLength = packetInfo.packet()->getLength() - rtpHeader->headerLength();
-
-    const auto mixerAudioBufferItr = _mixerSsrcAudioBuffers.find(ssrc.get());
-    if (mixerAudioBufferItr == _mixerSsrcAudioBuffers.cend())
-    {
-        logger::debug("New ssrc %u seen, sequence %u, sending request to add audio buffer",
-            _loggableId.c_str(),
-            ssrc.get(),
-            sequenceNumber.get());
-        _mixerSsrcAudioBuffers.emplace(ssrc, nullptr);
-        _messageListener.asyncAllocateAudioBuffer(*this, ssrc.get());
-    }
-    else if (!mixerAudioBufferItr->second)
-    {
-        logger::debug("new ssrc %u seen again, sequence %u, audio buffer is already requested",
-            _loggableId.c_str(),
-            ssrc.get(),
-            sequenceNumber.get());
-    }
-    else
-    {
-        const auto samples = payloadLength / bytesPerSample;
-        if (!mixerAudioBufferItr->second->write(reinterpret_cast<int16_t*>(payloadStart), samples))
-        {
-            if (!overrunLogSpamGuard)
-            {
-                logger::debug("Failed to write packet, buffer overrun, ssrc %u, sequence %u",
-                    _loggableId.c_str(),
-                    ssrc.get(),
-                    sequenceNumber.get());
-            }
-            overrunLogSpamGuard = true;
-        }
-    }
-}
-
 void EngineMixer::processIncomingRtpPackets(const uint64_t timestamp)
 {
     uint32_t numRtpPackets = 0;
@@ -2005,139 +1025,6 @@ void EngineMixer::processIncomingRtpPackets(const uint64_t timestamp)
     else
     {
         _lastReceiveTime = timestamp;
-    }
-}
-
-void EngineMixer::forwardVideoRtpPacket(IncomingPacketInfo& packetInfo, const uint64_t timestamp)
-{
-    auto rtpHeader = rtp::RtpHeader::fromPacket(*packetInfo.packet());
-    if (!rtpHeader)
-    {
-        assert(false); // this should have been checked multiple times by now. Transport, ReceiveJob, RtxReceiveJob
-        return;
-    }
-
-    const auto senderEndpointIdHash = packetInfo.packet()->endpointIdHash;
-
-    _lastVideoPacketProcessed = timestamp;
-
-    for (auto& videoStreamEntry : _engineVideoStreams)
-    {
-        const auto endpointIdHash = videoStreamEntry.first;
-        auto videoStream = videoStreamEntry.second;
-        if (!videoStream)
-        {
-            continue;
-        }
-
-        if (&videoStream->transport == packetInfo.transport())
-        {
-            continue;
-        }
-
-        if (!_engineStreamDirector->shouldForwardSsrc(endpointIdHash, packetInfo.inboundContext()->ssrc))
-        {
-            auto* senderVideoStream = _engineVideoStreams.getItem(senderEndpointIdHash);
-            if (senderVideoStream)
-            {
-                bool fwdSomeLevel = false;
-                for (auto& level : senderVideoStream->simulcastStream.getLevels())
-                {
-                    fwdSomeLevel |= _engineStreamDirector->shouldForwardSsrc(endpointIdHash, level.ssrc);
-                }
-                if (!fwdSomeLevel)
-                {
-                    logger::debug("no video is forwarded from %zu to %zu",
-                        _loggableId.c_str(),
-                        senderEndpointIdHash,
-                        endpointIdHash);
-                }
-            }
-            continue;
-        }
-
-        if (shouldSkipBecauseOfWhitelist(*videoStream, packetInfo.inboundContext()->ssrc))
-        {
-            continue;
-        }
-
-        uint32_t ssrc = 0;
-        const bool ssrcRewrite = videoStream->ssrcRewrite;
-        SsrcOutboundContext* ssrcOutboundContext = nullptr;
-        if (ssrcRewrite)
-        {
-            const auto& screenShareSsrcMapping = _activeMediaList->getVideoScreenShareSsrcMapping();
-            if (screenShareSsrcMapping.isSet() && screenShareSsrcMapping.get().first == senderEndpointIdHash &&
-                screenShareSsrcMapping.get().second.ssrc == packetInfo.inboundContext()->ssrc)
-            {
-                ssrc = screenShareSsrcMapping.get().second.rewriteSsrc;
-            }
-            else if (_engineStreamDirector->getPinTarget(endpointIdHash) == senderEndpointIdHash &&
-                !_activeMediaList->isInUserActiveVideoList(senderEndpointIdHash))
-            {
-                if (videoStream->pinSsrc.isSet())
-                {
-                    ssrc = videoStream->pinSsrc.get().ssrc;
-                }
-                else
-                {
-                    assert(false);
-                    continue;
-                }
-            }
-            else
-            {
-                const auto& videoSsrcRewriteMap = _activeMediaList->getVideoSsrcRewriteMap();
-                const auto* rewriteMap = videoSsrcRewriteMap.getItem(senderEndpointIdHash);
-                if (!rewriteMap)
-                {
-                    continue;
-                }
-                ssrc = (*rewriteMap)[0].main;
-            }
-
-            ssrcOutboundContext = obtainOutboundSsrcContext(videoStream->endpointIdHash,
-                videoStream->ssrcOutboundContexts,
-                ssrc,
-                videoStream->rtpMap);
-        }
-        else
-        {
-            // non rewrite recipients gets the video sent on same ssrc as level 0.
-            ssrc = packetInfo.inboundContext()->defaultLevelSsrc;
-            ssrcOutboundContext = obtainOutboundForwardSsrcContext(videoStream->endpointIdHash,
-                videoStream->ssrcOutboundContexts,
-                ssrc,
-                videoStream->rtpMap);
-        }
-
-        if (!ssrcOutboundContext)
-        {
-            continue;
-        }
-
-        if (!videoStream->transport.isConnected())
-        {
-            continue;
-        }
-
-        ssrcOutboundContext->onRtpSent(timestamp); // marks that we have active jobs on this ssrc context
-        auto packet = memory::makeUniquePacket(_sendAllocator, *packetInfo.packet());
-        if (packet)
-        {
-            videoStream->transport.getJobQueue().addJob<VideoForwarderRewriteAndSendJob>(*ssrcOutboundContext,
-                *(packetInfo.inboundContext()),
-                std::move(packet),
-                videoStream->transport,
-                packetInfo.extendedSequenceNumber(),
-                _messageListener,
-                videoStream->endpointIdHash,
-                *this);
-        }
-        else
-        {
-            logger::warn("send allocator depleted FwdRewrite", _loggableId.c_str());
-        }
     }
 }
 
@@ -2302,120 +1189,6 @@ void EngineMixer::processIncomingTransportFbRtcpPacket(const transport::RtcTrans
             blp,
             timestamp,
             transport->getRtt());
-    }
-}
-
-void EngineMixer::mixSsrcBuffers()
-{
-    memset(_mixedData, 0, samplesPerIteration * codec::Opus::bytesPerSample);
-    for (auto& mixerAudioBufferEntry : _mixerSsrcAudioBuffers)
-    {
-        if (!mixerAudioBufferEntry.second)
-        {
-            continue;
-        }
-        if (mixerAudioBufferEntry.second->isPreBuffering())
-        {
-            continue;
-        }
-
-        if (mixerAudioBufferEntry.second->getLength() < samplesPerIteration)
-        {
-            logger::debug("mixerAudioBufferEntry underrun", _loggableId.c_str());
-            mixerAudioBufferEntry.second->setPreBuffering();
-            continue;
-        }
-        else if (mixerAudioBufferEntry.second->getLength() < minimumSamplesInBuffer)
-        {
-            mixerAudioBufferEntry.second->insertSilence(samplesPerIteration);
-        }
-
-        mixerAudioBufferEntry.second->addToMix(_mixedData, samplesPerIteration, mixSampleScaleFactor);
-    }
-}
-
-inline void EngineMixer::processAudioStreams()
-{
-    for (auto& audioStreamEntry : _engineAudioStreams)
-    {
-        auto audioStream = audioStreamEntry.second;
-        auto isContributingToMix = false;
-        AudioBuffer* audioBuffer = nullptr;
-
-        if (audioStream->remoteSsrc.isSet())
-        {
-            audioBuffer = _mixerSsrcAudioBuffers.getItem(audioStream->remoteSsrc.get());
-            isContributingToMix = audioBuffer && !audioBuffer->isPreBuffering();
-
-            if (!audioStream->audioMixed || !audioStream->transport.isConnected())
-            {
-                if (isContributingToMix)
-                {
-                    audioBuffer->drop(samplesPerIteration);
-                }
-                continue;
-            }
-        }
-
-        if (!audioStream->audioMixed)
-        {
-            continue;
-        }
-
-        auto audioPacket = memory::makeUniquePacket(_audioAllocator);
-        if (!audioPacket)
-        {
-            return;
-        }
-
-        auto rtpHeader = rtp::RtpHeader::create(*audioPacket);
-        rtpHeader->ssrc = audioStream->localSsrc;
-
-        auto payloadStart = rtpHeader->getPayload();
-        const auto headerLength = rtpHeader->headerLength();
-        audioPacket->setLength(headerLength + samplesPerIteration * bytesPerSample);
-        memcpy(payloadStart, _mixedData, samplesPerIteration * bytesPerSample);
-
-        if (isContributingToMix && audioBuffer)
-        {
-            audioBuffer->removeFromMix(reinterpret_cast<int16_t*>(payloadStart),
-                samplesPerIteration,
-                mixSampleScaleFactor);
-            audioBuffer->drop(samplesPerIteration);
-        }
-
-        if (!audioStream->neighbours.empty())
-        {
-            for (auto& stream : _engineAudioStreams)
-            {
-                auto& peerAudioStream = *stream.second;
-                if (peerAudioStream.remoteSsrc.isSet() && peerAudioStream.transport.isConnected() &&
-                    areNeighbours(audioStream->neighbours, peerAudioStream.neighbours))
-                {
-                    auto* neighbourAudioBuffer = _mixerSsrcAudioBuffers.getItem(peerAudioStream.remoteSsrc.get());
-                    if (!neighbourAudioBuffer || neighbourAudioBuffer->isPreBuffering())
-                    {
-                        continue;
-                    }
-                    neighbourAudioBuffer->removeFromMix(reinterpret_cast<int16_t*>(payloadStart),
-                        samplesPerIteration,
-                        mixSampleScaleFactor);
-                }
-            }
-        }
-
-        auto* ssrcContext = obtainOutboundSsrcContext(audioStream->endpointIdHash,
-            audioStream->ssrcOutboundContexts,
-            audioStream->localSsrc,
-            audioStream->rtpMap);
-
-        if (ssrcContext)
-        {
-            audioStream->transport.getJobQueue().addJob<EncodeJob>(std::move(audioPacket),
-                *ssrcContext,
-                audioStream->transport,
-                _rtpTimestampSource);
-        }
     }
 }
 
@@ -2642,104 +1415,6 @@ void EngineMixer::sendDominantSpeakerMessageToAll()
     }
 }
 
-// This method is called after a video stream has been reconfigured. That has led to streams being removed and then
-// added again to ActiveMediaList and EngineStreamDirector. If the ssrc is an active inbound ssrc, we will set the
-// stream state again in EngineStreamDirector.
-void EngineMixer::restoreDirectorStreamActiveState(const EngineVideoStream& videoStream,
-    const SimulcastStream& simulcastStream)
-{
-    for (auto& simulcastLevel : simulcastStream.getLevels())
-    {
-        const auto ssrc = simulcastLevel.ssrc;
-        auto ssrcInboundContext = _ssrcInboundContexts.getItem(ssrc);
-        if (ssrcInboundContext)
-        {
-            ssrcInboundContext->makeReady();
-            _engineStreamDirector->streamActiveStateChanged(videoStream.endpointIdHash, ssrc, false);
-        }
-    }
-}
-
-void EngineMixer::markAssociatedAudioOutboundContextsForDeletion(const EngineAudioStream* senderAudioStream)
-{
-    if (!senderAudioStream->remoteSsrc.isSet())
-    {
-        return;
-    }
-
-    for (auto& audioStreamEntry : _engineAudioStreams)
-    {
-        auto* audioStream = audioStreamEntry.second;
-        if (audioStream == senderAudioStream || audioStream->ssrcRewrite)
-        {
-            // If we use ssrc rewrite there is no need to remove the outbound context as it is not there
-            continue;
-        }
-        const auto endpointIdHash = audioStreamEntry.first;
-
-        auto outboundContextItr = audioStream->ssrcOutboundContexts.find(senderAudioStream->remoteSsrc.get());
-        if (outboundContextItr != audioStream->ssrcOutboundContexts.end())
-        {
-            outboundContextItr->second.markedForDeletion = true;
-            logger::info("Marking unused audio outbound context for deletion, ssrc %u, endpointIdHash %lu",
-                _loggableId.c_str(),
-                outboundContextItr->first,
-                endpointIdHash);
-
-            static constexpr uint32_t feedbackSsrc = 0;
-            audioStream->transport.getJobQueue().addJob<FinalizeNonSsrcRewriteOutboundContextJob>(*this,
-                audioStream->transport,
-                outboundContextItr->second,
-                _engineSyncContext,
-                _messageListener,
-                feedbackSsrc);
-        }
-    }
-}
-
-void EngineMixer::markAssociatedVideoOutboundContextsForDeletion(const EngineVideoStream* senderVideoStream,
-    const uint32_t ssrc,
-    const uint32_t feedbackSsrc)
-{
-    for (auto& videoStreamEntry : _engineVideoStreams)
-    {
-        auto* videoStream = videoStreamEntry.second;
-        if (videoStream == senderVideoStream || videoStream->ssrcRewrite)
-        {
-            // If we use ssrc rewrite there is no need to remove the outbound context as it is not there
-            continue;
-        }
-        const auto endpointIdHash = videoStreamEntry.first;
-
-        auto* outboundContext = videoStream->ssrcOutboundContexts.getItem(ssrc);
-        if (outboundContext)
-        {
-            outboundContext->markedForDeletion = true;
-            logger::info("Marking unused video outbound context for deletion, ssrc %u, endpointIdHash %lu",
-                _loggableId.c_str(),
-                ssrc,
-                endpointIdHash);
-
-            videoStream->transport.getJobQueue().addJob<FinalizeNonSsrcRewriteOutboundContextJob>(*this,
-                videoStream->transport,
-                *outboundContext,
-                _engineSyncContext,
-                _messageListener,
-                feedbackSsrc);
-        }
-
-        outboundContext = videoStream->ssrcOutboundContexts.getItem(feedbackSsrc);
-        if (outboundContext)
-        {
-            outboundContext->markedForDeletion = true;
-            logger::info("Marking unused video outbound context for deletion, feedback ssrc %u, endpointIdHash %lu´",
-                _loggableId.c_str(),
-                ssrc,
-                endpointIdHash);
-        }
-    }
-}
-
 void EngineMixer::decommissionInboundContext(const uint32_t ssrc)
 {
     auto* ssrcContext = _ssrcInboundContexts.getItem(ssrc);
@@ -2748,18 +1423,6 @@ void EngineMixer::decommissionInboundContext(const uint32_t ssrc)
         _ssrcInboundContexts.erase(ssrc); // remove first or internalRemoveInboundSsrc may assert
         ssrcContext->sender->postOnQueue(utils::bind(&EngineMixer::internalRemoveInboundSsrc, this, ssrc));
         logger::info("Decommissioned inbound ssrc context %u", _loggableId.c_str(), ssrc);
-    }
-}
-
-void EngineMixer::processEngineMissingPackets(bridge::SsrcInboundContext& ssrcInboundContext)
-{
-    auto* videoStream = _engineVideoStreams.getItem(ssrcInboundContext.sender->getEndpointIdHash());
-    if (videoStream)
-    {
-        videoStream->transport.getJobQueue().addJob<bridge::ProcessMissingVideoPacketsJob>(ssrcInboundContext,
-            videoStream->localSsrc,
-            videoStream->transport,
-            _sendAllocator);
     }
 }
 
@@ -2778,154 +1441,6 @@ uint32_t EngineMixer::getMinRemoteClientDownlinkBandwidth() const
         minBitrate = std::min(minBitrate, itBb.second->minClientDownlinkBandwidth);
     }
     return minBitrate;
-}
-
-void EngineMixer::checkVideoBandwidth(const uint64_t timestamp)
-{
-    const bool isTimeToCheck = utils::Time::diffGE(_lastVideoBandwidthCheck, timestamp, utils::Time::sec * 3);
-    const bool shouldExecute = isTimeToCheck || _engineStreamDirector->needsSlidesBitrateAllocation();
-    if (!shouldExecute)
-    {
-        return;
-    }
-
-    _lastVideoBandwidthCheck = timestamp;
-    uint32_t minUplinkEstimate = 10000000;
-    bridge::SimulcastLevel* presenterSimulcastLevel = nullptr;
-    bridge::EngineVideoStream* presenterStream = nullptr;
-    for (auto videoIt : _engineVideoStreams)
-    {
-        auto& videoStream = *videoIt.second;
-        if (videoStream.simulcastStream.contentType == SimulcastStream::VideoContentType::SLIDES)
-        {
-            presenterSimulcastLevel = &videoStream.simulcastStream.levels[0];
-            presenterStream = videoIt.second;
-        }
-        else if (videoStream.secondarySimulcastStream.isSet() &&
-            videoStream.secondarySimulcastStream.get().contentType == SimulcastStream::VideoContentType::SLIDES)
-        {
-            presenterSimulcastLevel = &videoStream.secondarySimulcastStream.get().levels[0];
-            presenterStream = videoIt.second;
-        }
-        else
-        {
-            minUplinkEstimate = std::min(minUplinkEstimate, videoIt.second->transport.getUplinkEstimateKbps());
-        }
-    }
-
-    minUplinkEstimate = std::max(minUplinkEstimate, _config.slides.minBitrate.get());
-    if (needToUpdateMinUplinkEstimate(minUplinkEstimate, _minUplinkEstimate))
-    {
-        reportMinRemoteClientDownlinkBandwidthToBarbells(minUplinkEstimate);
-    }
-
-    minUplinkEstimate =
-        std::max(_config.slides.minBitrate.get(), std::min(minUplinkEstimate, getMinRemoteClientDownlinkBandwidth()));
-
-    // Local presenter. Sending TMBBR to restrict slides.
-    if (presenterStream)
-    {
-        const uint32_t slidesLimit = minUplinkEstimate * _config.slides.allocFactor;
-
-        logger::info("limiting bitrate for ssrc %u, at %u",
-            _loggableId.c_str(),
-            presenterSimulcastLevel->ssrc,
-            slidesLimit);
-
-        presenterStream->transport.getJobQueue().addJob<SetMaxMediaBitrateJob>(presenterStream->transport,
-            presenterStream->localSsrc,
-            presenterSimulcastLevel->ssrc,
-            slidesLimit,
-            _sendAllocator);
-
-        _engineStreamDirector->setSlidesSsrcAndBitrate(presenterSimulcastLevel->ssrc, slidesLimit);
-    }
-    else
-    {
-        _engineStreamDirector->setSlidesSsrcAndBitrate(0, 0);
-    }
-
-    // Any presenter (local or remote). Sending TMBBR to restrict video.
-    const bool slidesPresent = _activeMediaList->getVideoScreenShareSsrcMapping().isSet();
-
-    for (auto videoIt : _engineVideoStreams)
-    {
-        auto& videoStream = *videoIt.second;
-        const uint32_t highQualityBitrate = slidesPresent ? 750 : 8000;
-
-        if (3 == videoStream.simulcastStream.numLevels)
-        {
-            const auto& highResSsrc = videoStream.simulcastStream.levels[2].ssrc;
-            logger::info("setting high-res bitrate for ssrc %u, at %u",
-                _loggableId.c_str(),
-                highResSsrc,
-                highQualityBitrate);
-
-            videoStream.transport.getJobQueue().addJob<SetMaxMediaBitrateJob>(videoStream.transport,
-                videoStream.localSsrc,
-                highResSsrc,
-                highQualityBitrate,
-                _sendAllocator);
-
-            // WebRTC's incorrect implementation of TMBBR leads assigned bitrate to be
-            // used for all members of the simulcast group, instead of the single layer.
-            // That, in turn, leads to high res layer being switched on sometimes for a
-            // short burst of frames. SFU detects constant switching on and off as "unstable"
-            // behaviour and disables such layer forever. To prevent that side effect, we
-            // reset "inactiveTransitionCount" when switching slides off.
-            if (_slidesPresent && !slidesPresent)
-            {
-                auto inboundSsrcContext = _ssrcInboundContexts.getItem(highResSsrc);
-                if (inboundSsrcContext)
-                {
-                    inboundSsrcContext->makeReady();
-                }
-            }
-        }
-    }
-    _slidesPresent = slidesPresent;
-}
-
-void EngineMixer::startProbingVideoStream(EngineVideoStream& engineVideoStream)
-{
-    if (engineVideoStream.feedbackRtpMap.isEmpty())
-    {
-        return;
-    }
-
-    auto* outboundContext = obtainOutboundSsrcContext(engineVideoStream.endpointIdHash,
-        engineVideoStream.ssrcOutboundContexts,
-        engineVideoStream.localSsrc,
-        engineVideoStream.feedbackRtpMap);
-
-    if (outboundContext)
-    {
-        engineVideoStream.transport.postOnQueue(utils::bind(&transport::RtcTransport::setRtxProbeSource,
-            &engineVideoStream.transport,
-            engineVideoStream.localSsrc,
-            &outboundContext->rewrite.lastSent.sequenceNumber,
-            outboundContext->rtpMap.payloadType));
-    }
-    _probingVideoStreams = true;
-}
-
-void EngineMixer::stopProbingVideoStream(const EngineVideoStream& engineVideoStream)
-{
-    if (engineVideoStream.feedbackRtpMap.isEmpty())
-    {
-        return;
-    }
-
-    engineVideoStream.transport.postOnQueue(utils::bind(&transport::RtcTransport::setRtxProbeSource,
-        &engineVideoStream.transport,
-        engineVideoStream.localSsrc,
-        nullptr,
-        engineVideoStream.feedbackRtpMap.payloadType));
-}
-
-bool EngineMixer::isVideoInUse(const uint64_t timestamp, const uint64_t threshold) const
-{
-    return utils::Time::diffLE(_lastVideoPacketProcessed, timestamp, threshold);
 }
 
 void EngineMixer::checkIfRateControlIsNeeded(const uint64_t timestamp)
@@ -3036,49 +1551,9 @@ void EngineMixer::mapSsrc2UserId(uint32_t ssrc, uint32_t usid)
     _audioSsrcToUserIdMap.emplace(ssrc, usid);
 }
 
-bool EngineMixer::asyncReconfigureVideoStream(const transport::RtcTransport& transport,
-    const SsrcWhitelist& ssrcWhitelist,
-    const SimulcastStream& simulcastStream,
-    const utils::Optional<SimulcastStream>& secondarySimulcastStream)
-{
-    return post(utils::bind(&EngineMixer::reconfigureVideoStream,
-        this,
-        std::ref(transport),
-        ssrcWhitelist,
-        simulcastStream,
-        secondarySimulcastStream));
-}
-
-bool EngineMixer::asyncAddAudioBuffer(const uint32_t ssrc, AudioBuffer* audioBuffer)
-{
-    return post(utils::bind(&EngineMixer::addAudioBuffer, this, ssrc, audioBuffer));
-}
-
-bool EngineMixer::asyncRemoveStream(const EngineAudioStream* engineAudioStream)
-{
-    return post([=]() { this->removeStream(engineAudioStream); });
-}
-
-bool EngineMixer::asyncRemoveStream(const EngineVideoStream* stream)
-{
-    return post([=]() { this->removeStream(stream); });
-}
-
 bool EngineMixer::asyncRemoveStream(const EngineDataStream* stream)
 {
     return post([=]() { this->removeStream(stream); });
-}
-
-bool EngineMixer::asyncAddVideoPacketCache(const uint32_t ssrc,
-    const size_t endpointIdHash,
-    PacketCache* videoPacketCache)
-{
-    return post(utils::bind(&EngineMixer::addVideoPacketCache, this, ssrc, endpointIdHash, videoPacketCache));
-}
-
-bool EngineMixer::asyncReconfigureAudioStream(const transport::RtcTransport& transport, const uint32_t remoteSsrc)
-{
-    return post(utils::bind(&EngineMixer::reconfigureAudioStream, this, std::cref(transport), remoteSsrc));
 }
 
 bool EngineMixer::asyncStartTransport(transport::RtcTransport& transport)
@@ -3086,35 +1561,9 @@ bool EngineMixer::asyncStartTransport(transport::RtcTransport& transport)
     return post(utils::bind(&EngineMixer::startTransport, this, std::ref(transport)));
 }
 
-bool EngineMixer::asyncAddAudioStream(EngineAudioStream* engineAudioStream)
-{
-    return post(utils::bind(&EngineMixer::addAudioStream, this, engineAudioStream));
-}
-
-bool EngineMixer::asyncAddVideoStream(EngineVideoStream* engineVideoStream)
-{
-    return post(utils::bind(&EngineMixer::addVideoStream, this, engineVideoStream));
-}
-
 bool EngineMixer::asyncAddDataSteam(EngineDataStream* engineDataStream)
 {
     return post(utils::bind(&EngineMixer::addDataSteam, this, engineDataStream));
-}
-
-bool EngineMixer::asyncPinEndpoint(const size_t endpointIdHash, const size_t targetEndpointIdHash)
-{
-    return post(utils::bind(&EngineMixer::pinEndpoint, this, endpointIdHash, targetEndpointIdHash));
-}
-
-bool EngineMixer::asyncSendEndpointMessage(const size_t toEndpointIdHash,
-    const size_t fromEndpointIdHash,
-    memory::UniqueAudioPacket& packet)
-{
-    return post(utils::bind(&EngineMixer::sendEndpointMessage,
-        this,
-        toEndpointIdHash,
-        fromEndpointIdHash,
-        utils::moveParam(packet)));
 }
 
 bool EngineMixer::asyncHandleSctpControl(const size_t endpointIdHash, memory::UniquePacket& packet)
